Update ssl.cr
changed HostDataSSLCert#serial type to a UInt128. This supports bigger serial numbers and fixes any parsing issues with numbers larger than an Int64 supports

+++ src/shodan/parsers/host_services/ssl.cr
@@ -68,7 +68,7 @@ module Shodan
     property version : Int32?
     property extensions : Array(HostDataSSLExtensions)?
     property fingerprint : Hash(String, String)?
-    property serial : Int64?
+    property serial : UInt128?
     property subject : HostDataSslSubject
     property issuer : HostDataSslIssuer
   end
